Removed duplicate/depracated code for rich mode, consolidated into single message processing service
